Widgets // Move profile-based desktop widgets into PZWidgetsKit.
