+++ WidgetExtension/Modules/DesktopWidgets/DualProfileWidget.swift
@@ -20,213 +20,11 @@ struct DualProfileWidget: Widget {
             intent: SelectDualProfileIntent.self,
             provider: DualProfileWidgetProvider()
         ) { entry in
-            DualProfileWidgetView(entry: entry)
+            DualProfileWidgetView<WidgetRefreshIntent>(entry: entry)
         }
         .configurationDisplayName("pzWidgetsKit.statusDualProfile.title".i18nWidgets)
         .description("pzWidgetsKit.statusDualProfile.enquiry.title".i18nWidgets)
         .supportedFamilies([.systemMedium, .systemLarge, .systemExtraLarge])
         .containerBackgroundRemovable(false)
-    }
-}
-
-// MARK: - DualProfileWidgetView
-
-@available(watchOS, unavailable)
-private struct DualProfileWidgetView: View {
-    // MARK: Internal
-
-    @Environment(\.widgetFamily) var family: WidgetFamily
-
-    let entry: DualProfileWidgetProvider.Entry
-
-    var resultSlot1: Result<any DailyNoteProtocol, any Error> { entry.resultSlot1 }
-    var resultSlot2: Result<any DailyNoteProtocol, any Error> { entry.resultSlot2 }
-    var viewConfig: WidgetViewConfig { entry.viewConfig }
-
-    var subEntry1: MainWidgetProvider.Entry {
-        .init(
-            date: entry.date,
-            result: entry.resultSlot1,
-            viewConfig: entry.viewConfig,
-            profile: entry.profileSlot1,
-            pilotAssetMap: entry.pilotAssetMap,
-            events: entry.events.filter { $0.game == entry.profileSlot1?.game }
-        )
-    }
-
-    var subEntry2: MainWidgetProvider.Entry {
-        .init(
-            date: entry.date,
-            result: entry.resultSlot2,
-            viewConfig: entry.viewConfig,
-            profile: entry.profileSlot2,
-            pilotAssetMap: entry.pilotAssetMap,
-            events: entry.events.filter { $0.game == entry.profileSlot2?.game }
-        )
-    }
-
-    var widgetFamilyForComponents: WidgetFamily {
-        switch family {
-        case .systemSmall: .systemSmall
-        case .systemMedium: .systemSmall
-        case .systemLarge: .systemMedium
-        case .systemExtraLarge: .systemLarge
-        default: family
-        }
-    }
-
-    var body: some View {
-        Group {
-            switch family {
-            case .systemSmall: EmptyView() // Not supported.
-            case .systemMedium:
-                HStack {
-                    contents
-                }
-                .frame(maxWidth: .infinity)
-                .overlay {
-                    overlayDivider(isVertical: false)
-                }
-            case .systemLarge:
-                VStack {
-                    contents
-                }
-                .frame(maxWidth: .infinity)
-                .overlay {
-                    overlayDivider(isVertical: true)
-                }
-            case .systemExtraLarge:
-                HStack {
-                    VStack {
-                        contents
-                    }
-                    .frame(maxWidth: .infinity)
-                    .overlay {
-                        overlayDivider(isVertical: true)
-                    }
-                    if family == .systemExtraLarge {
-                        officialFeedBlock()
-                            .frame(maxWidth: 300)
-                    }
-                }
-            case .accessoryCorner: EmptyView()
-            case .accessoryCircular: EmptyView()
-            case .accessoryRectangular: EmptyView()
-            case .accessoryInline: EmptyView()
-            @unknown default: EmptyView()
-            }
-        }
-        .padding()
-        .environment(\.colorScheme, .dark)
-        .pzWidgetContainerBackground(viewConfig: viewConfig)
-    }
-
-    @ViewBuilder var contents: some View {
-        if !viewConfig.useTinyGlassDisplayStyle { Spacer(minLength: 0) }
-        drawSingleEntry(subEntry1)
-        Spacer(minLength: 15)
-        drawSingleEntry(subEntry2)
-        if !viewConfig.useTinyGlassDisplayStyle { Spacer(minLength: 0) }
-    }
-
-    // MARK: Private
-
-    @ViewBuilder
-    private func officialFeedBlock() -> some View {
-        VStack(alignment: .trailing) {
-            let officialFeedList = OfficialFeedList4WidgetsView(
-                events: entry.events,
-                showLeadingBorder: false,
-                refreshIntent: WidgetRefreshIntent()
-            )
-            .contentShape(.rect)
-            switch viewConfig.useTinyGlassDisplayStyle {
-            case false:
-                officialFeedList
-                    .padding(.leading, 14)
-                Spacer()
-                WeekdayDisplayView()
-                    .frame(maxWidth: .infinity, alignment: .trailing)
-            case true:
-                OfficialFeedList4WidgetsView(
-                    events: entry.events,
-                    showLeadingBorder: false,
-                    refreshIntent: WidgetRefreshIntent()
-                )
-                .padding(.horizontal, 10)
-                .padding(.vertical, 10)
-                .widgetAccessibilityBackground(enabled: viewConfig.useTinyGlassDisplayStyle)
-                Spacer()
-                WeekdayDisplayView()
-                    .padding(.horizontal, 10)
-                    .widgetAccessibilityBackground(enabled: viewConfig.useTinyGlassDisplayStyle)
-                    .frame(maxWidth: .infinity, alignment: .trailing)
-            }
-        }
-        .frame(maxWidth: .infinity)
-    }
-
-    @ViewBuilder
-    private func drawSingleEntry(_ givenEntry: MainWidgetProvider.Entry) -> some View {
-        Group {
-            switch givenEntry.result {
-            case let .success(dailyNote):
-                let profileName = givenEntry.profile?.name
-                switch family {
-                case .systemMedium, .systemSmall:
-                    MainInfo(
-                        entry: givenEntry,
-                        dailyNote: dailyNote,
-                        viewConfig: viewConfig,
-                        accountName: profileName
-                    )
-                // case .systemLarge, .systemExtraLarge:
-                default:
-                    switch viewConfig.expeditionDisplayPolicy {
-                    case .neverDisplay:
-                        Color.clear
-                    case .displayWhenAvailable:
-                        MainInfoWithDetail(
-                            entry: givenEntry,
-                            dailyNote: dailyNote,
-                            viewConfig: viewConfig,
-                            accountName: profileName
-                        )
-                    case .displayExclusively:
-                        MainInfoWithExpedition(
-                            entry: givenEntry,
-                            dailyNote: dailyNote,
-                            viewConfig: viewConfig,
-                            accountName: profileName
-                        )
-                    }
-                }
-            case let .failure(error):
-                WidgetErrorView(
-                    error: error,
-                    message: viewConfig.noticeMessage ?? "",
-                    refreshIntent: WidgetRefreshIntent(
-                        dailyNoteUIDWithGame: givenEntry.profile?.uidWithGame
-                    )
-                )
-            }
-        }
-    }
-
-    @ViewBuilder
-    private func overlayDivider(isVertical: Bool) -> some View {
-        let dividerViewRAW = Group {
-            Color.clear
-            Divider().overlay {
-                Color.black
-            }
-            .blendMode(.colorDodge)
-            .padding()
-            Color.clear
-        }
-        switch isVertical {
-        case true: VStack { dividerViewRAW }
-        case false: HStack { dividerViewRAW }
-        }
     }
 }
