+++ WidgetExtension/Modules/DesktopWidgets/MainWidget.swift
@@ -20,49 +20,11 @@ struct MainWidget: Widget {
             intent: SelectAccountIntent.self,
             provider: MainWidgetProvider()
         ) { entry in
-            WidgetViewEntryView(entry: entry, noBackground: false)
+            SingleProfileWidgetView<WidgetRefreshIntent>(entry: entry, noBackground: false)
         }
         .configurationDisplayName("pzWidgetsKit.status.title".i18nWidgets)
         .description("pzWidgetsKit.status.enquiry.title".i18nWidgets)
         .supportedFamilies([.systemSmall, .systemMedium, .systemLarge, .systemExtraLarge])
         .containerBackgroundRemovable(false)
-    }
-}
-
-// MARK: - WidgetViewEntryView
-
-@available(watchOS, unavailable)
-struct WidgetViewEntryView: View {
-    @Environment(\.widgetFamily) var family: WidgetFamily
-
-    let entry: MainWidgetProvider.Entry
-    let noBackground: Bool
-
-    var result: Result<any DailyNoteProtocol, any Error> { entry.result }
-    var viewConfig: WidgetViewConfig { entry.viewConfig }
-    var accountName: String? { entry.profile?.name }
-
-    var body: some View {
-        ZStack {
-            switch result {
-            case let .success(dailyNote):
-                WidgetMainView(
-                    entry: entry,
-                    dailyNote: dailyNote,
-                    viewConfig: viewConfig,
-                    accountName: accountName
-                )
-            case let .failure(error):
-                WidgetErrorView(
-                    error: error,
-                    message: viewConfig.noticeMessage ?? "",
-                    refreshIntent: WidgetRefreshIntent(
-                        dailyNoteUIDWithGame: entry.profile?.uidWithGame
-                    )
-                )
-            }
-        }
-        .environment(\.colorScheme, .dark)
-        .pzWidgetContainerBackground(viewConfig: noBackground ? nil : viewConfig)
     }
 }
